Refactor main.ts to include clone, run, and stop VM functions
