+++ apps/vm_runner/base_url.ts
@@ -1,0 +1,1 @@
+export const baseUrl = "http://localhost:8080";
